pyproject.toml
[tool.poetry]
name = "djangoproject1"
version = "0.1.0"
description = ""
authors = ["Andrey Matveev <a.matveev@idaproject.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
django-allauth = "^0.52.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
